mlops-experiments/mlops-core/model/requirements.txt
numpy==2.3.3
mlflow
scikit-learn
python-dotenv
boto3
prometheus_client
